+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/ya.make
@@ -4,6 +4,7 @@ SRCS(
     fetched_data.cpp
     columns_set.cpp
     iterator.cpp
+    context.cpp
 )
 
 PEERDIR(
